building the foundation for pdfr
rebasing with doc fixing pr
