+++ Cargo.toml
@@ -2,7 +2,8 @@
 members = [
     "crates/core",
     "crates/markdownify",
-    "crates/rasteroid"
+    "crates/rasteroid",
+    "crates/pdfr"
 ]
 resolver = "3"
