+++ crates/pdfr/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "markdownify"
-version = "0.1.1"
+name = "pdfr"
+version = "0.1.0"
 authors = ["Meron Bossin"]
-description = "markitdown in rust - convert various document and files into markdown"
-keywords = ["conversion", "parser", "markitdown", "document", "markdown"]
+description = "high level pdf parser"
+keywords = ["pdf", "parser", "high-level", "reader", "pdf2"]
 categories = ["parser-implementations", "text-processing"]
 repository = "https://github.com/Skardyy/mcat"
 homepage = "https://github.com/Skardyy/mcat"
@@ -13,8 +13,4 @@ readme = "README.md"
 edition = "2024"
 
 [dependencies]
-tempfile = "3.19.1"
-csv = "1.3.1"
-zip = "2.6.1"
-calamine = "0.25.0"
-quick-xml = "0.37.4"
+lopdf = "0.36.0"
