+++ crates/rasteroid/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.1"
 authors = ["Meron Bossin"]
 description = "turn images / videos into inline content for you terminal (iterm / kitty / sixel)"
 keywords = ["sixel", "iterm", "kitty", "terminal", "terminal-graphics"]
-categories = ["graphics", "rendering"]
+categories = ["graphics", "rendering", "command-line-utilities"]
 repository = "https://github.com/Skardyy/mcat"
 homepage = "https://github.com/Skardyy/mcat"
 documentation = "https://github.com/Skardyy/mcat"
